requirements.txt
numpy
scipy
jupyter
matplotlib
Pillow
scikit-learn
cython
tensorflow-gpu
